api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
pgvector==0.3.6
sentence-transformers==3.3.1
pandas==2.2.3
python-dotenv==1.0.1
